pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trackzero"
version = "0.0.1"
description = "Learning whole-body tracking policies from physics alone"
requires-python = ">=3.10"
dependencies = [
    "mujoco>=3.4.0",
    "torch>=2.5.1",
    "numpy>=1.26",
    "matplotlib>=3.10",
    "h5py",
    "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-timeout"]

[tool.setuptools.packages.find]
include = ["trackzero*"]
